requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
openpyxl==3.1.2
python-decouple==3.8
django-cors-headers==4.3.1
dj-database-url==2.1.0
python-dateutil==2.8.2
